More & More
I forgot it was on. So terminal in the background made 1038 poems.

The one made as i returned to its still active working thread read:

 "He put on his boot scene of town.
 His face becomes one tone…”

and the one after

“I want to control. Oh, no face….”

I watched it for a few moments being eloquent and more clever than i
can be and inexhaustible, and I grew tired. Who is this machine made
for? This human. This mind. This matter.

“ The eyes of someone whisper
 crossing life's world.
 Forgive it “

Finally:

“And rustle in the shack
 Of horseplay. Flap. chemise, and the verrai-je ealdgesias,
 She took me by the Marble hallway, without a toxins,
 And called a rope up on a bench bared my head,
 The back year cup; and then again went aside,
 Went straight down to wake. “

		~ + ~